Mead
This gem is for easily adding honeypots to forms as well as being able
to easily obfuscate form fields.

It adds tag helpers for forms `mead_honeypot_tag` and
`mead_obfuscate_tag` as well as form helper methods `mead_honeypot` and
`mead_obfuscate` which themselves can be used from a form builder.

This makes it easy to not only add honeypots and mask params but to make
them blend in with the rest of your form to better thwart targeted bots.

You can easily check for honeypots by calling `honeypot_present?` and
unmask your params via `mead_params`.