Update param deobfuscation to be more natural
In order to make the accessing of masked params this commit updates the
mead_params method so that instead of accessing very specific sections
of your params it will instead recursively check all of the params
(ActionController::Parameters) for keys that are obfuscated, deobfuscate
them, and then finally returns a new ActionController::Parameters object
with the cleaned up params.

Functionally anyone should be able to use mead_params the way they would
params.