Fix Load Order
In order to make sure we get the configuration information before
attaching to the ActionController we need to actually go through the
configuration process. This means that we'll no longer be auto-loaded
going forward but that is probalby for the best.

Now users will either have to define a mead_captcha.rb file in the
initializers or they will need to manually include MeadCaptcha in their
controller they wish to protect. This should help us avoid issues where
controller actions we don't care about aren't 'proteted'.